Use `MonoidMap` to simplify `Inventory` functions. (#2332)
* Add dependency on `monoidmap` package.

This package provides the `MonoidMap` data type, a mapping from keys to
monoidal values.

We also add a dependency on `monoidmap-aeson`, which provides `ToJSON`
and `FromJSON` instances for `MonoidMap`.

These dependencies will be used in later commits.

* Redefine `Inventory.byName` to use `MonoidMap`.

The original definition of `Inventory` defines `byName` in terms of `Map`:

```hs
data Inventory = Inventory
  { counts :: IntMap (Count, Entity)
  , byName :: Map Text IntSet
  , inventoryHash :: Int
  }
```

The use of `Map` makes it possible to represent certain states in two
different ways.

Consider a situation where a given name `name` does not map to any
hashes. The use of `Map` means that we can represent this singular state
in two distinct ways:

1. The `name` is **not** present in the `Map`, and `lookup` returns `Nothing`.
2. The `name` **is** present in the `Map`, but maps to `Just IntSet.empty`.

The current code treats these two representations identically, but is
forced to handle both cases explicitly.

For example, consider the `lookupByName` function:

```hs
lookupByName name (Inventory cs byN _) =
  maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (M.lookup (T.toLower name) byN)
```

When `name` is **not** present in the map, `lookupByName` evaluates to `[]`:

```hs
   maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (M.lookup (T.toLower name) byN)
== maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (Nothing)
==       []
```

When `name` **is** present in the map, but maps to the empty `IntSet`,
`lookupByName` also evaluates to `[]`

```hs
   maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (M.lookup (T.toLower name) byN)
== maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (Just IntSet.empty)
==          (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (     IntSet.empty)
==           mapMaybe (fmap snd . (cs IM.!?))  (IS.elems        IntSet.empty)
==           mapMaybe (fmap snd . (cs IM.!?))  []
==                                             []
```

In contrast, the `MonoidMap` data type provides a single, _canonical_
representation for null or empty values, according to the following
conceptual mapping:

| `Map.lookup k m`        | ⟼ | `MonoidMap.get k m`     |
|:------------------------|---|:------------------------|
| `Nothing`               | ⟼ | `mempty`                |
| `Just v \| v == mempty` | ⟼ | `mempty`                |
| `Just v \| v /= mempty` | ⟼ | `v`                     |

In our context, `mempty` corresponds to `IntSet.empty`.

This commit makes the following transformation:

```patch
  data Inventory = Inventory
    { counts :: IntMap (Count, Entity)
-   , byName ::       Map Text IntSet
+   , byName :: MonoidMap Text IntSet
    , inventoryHash :: Int
    }
```

This allows us to simplify the definition of `lookupByName`:

```patch
- maybe [] (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (M.lookup (T.toLower name) byN)
+          (mapMaybe (fmap snd . (cs IM.!?)) . IS.elems) (MM.get   (T.toLower name) byN)
```

* Use `MonoidMap.union` to simplify `Inventory.union`.

This relies on the following equivalence:

```hs
MonoidMap.union
    == MonoidMap.unionWith (<>)
    == MonoidMap.unionWith IntSet.union
```

* Use `IntMap.restrictKeys` to simplify `lookupByName`.

Instead of converting the set of hashes to a list and iterating over
them one by one to see whether they're in the count map, we directly
restrict the count map to only include mappings for the hashes we're
interested in.

* Use `IntSet.insert` to simplify `insertCount`.

This relies on the following equivalence:

```hs
IntSet.insert a as == IntSet.union (IntSet.singleton a) as
```