Use `MonoidMap` for `robotsByLocation`.
This commit redefines `Robots.robotsByLocation` to use a nested
`MonoidMap` instead of a nested `Map`:

```patch
  data Robots = Robots
    { ...
-   , _robotsByLocation ::       Map SubworldName (      Map Location IntSet)
+   , _robotsByLocation :: MonoidMap SubworldName (MonoidMap Location IntSet)
    , ...
    }
```

Queries of the original nested `Map` did not distinguish between
mappings to `Nothing` and mappings to `Just mempty`, whether in the
outer map or the inner map. By using `MonoidMap`, we can simplify such
queries, as well as functions that create or update the map.

For example, `robotsAtLocation` no longer needs to explicitly handle
cases where the outer map does not contain a given `SubworldName`, or
where the inner map does not contain a given `Location`:

```patch
  robotsAtLocation loc gs =
    mapMaybe (`IM.lookup` (gs ^. robotInfo . robotMap))
-     . maybe [] IS.toList
-     . M.lookup (loc ^. planar)
-     . M.findWithDefault mempty (loc ^. subworld)
+     . IS.toList
+     . MM.get (loc ^. planar)
+     . MM.get (loc ^. subworld)
      . view (robotInfo . robotsByLocation)
      $ gs
```

Insertion and removal can now be expressed in terms of simple nested
adjustments, avoiding the need to have special handling code for
mappings to null values. It's arguably now easier to see the inherent
symmetry of these operations:

```hs
addRobotToLocation rid rLoc =
  robotsByLocation %=
  MM.adjust (MM.adjust (IS.insert rid) (rLoc ^. planar)) (rLoc ^. subworld)

removeRobotFromLocation rid rLoc =
  robotsByLocation %=
  MM.adjust (MM.adjust (IS.delete rid) (rLoc ^. planar)) (rLoc ^. subworld)
```