Water disappears when placed, via new `evanescent` property (#2358)
Based on the [discussion here](https://github.com/swarm-game/swarm/issues/1435#issuecomment-2425192480).  The idea is that you can `grab` a bucket full of water, but if you try to `place` it somewhere you just dump it and it disappears --- it doesn't create a permanent lake.  This also solves the awkward issue that technically if you created a lake underneath yourself without a `boat` equipped you should drown.