Make bind tutorial easier to understand (#2361)
* move the bind explanation back to the first goal
  * remove the goal description of `Cmd Unit` as the previous types tutorial explained it already
* add a new hidden objective that will recognize that the sent robot has stopped short of delivering the artifact
  * make the "Start over" button the default on loss
* simplify room layout 
  * make the pedestal a golden connected boundary to make it easier to see
  * move the base entrance to the center so that it's easier to guess the number of moves
* closes #2310

| <img width="117" alt="Screenshot 2025-03-30 at 1 16 11 AM" src="https://github.com/user-attachments/assets/f3b29f00-5ac5-4378-903b-d0455acba3c7" /> |
| --- |
| New room layout |

The description of _bind notation_ was not shown unless the player sent a robot that successfully grabbed the artifact.
This could have paralyzed the player if they wanted to get it right the first time.

Now, the goal description is more upfront. If the player fails their one try, a hidden objective automatically detects it, fails the main objective, and a window with a retry option pops up.