Grant `AttemptSelfDestructBase` achievement for movement as well as `selfdestruct` (#2356)
Previously, we only granted this achievement when literally executing the `selfdestruct` command, but that's kind of difficult since you have to first craft a `detonator`.  This PR also grants the achievement from the `applyMoveFailureEffect` function, so e.g. if you attempt to move the base into water without a `boat` equipped.