toplevel records zoom (#2362)
Continues toward #2276.

This PR incrementally pushes the state scope to the narrower `PlayState` rather than `AppState` for many utility functions, re-simplifying the code in many cases where an extra `playState` lens layer was added in #2282.  Counting the occurrences of the `playState` lens has been a good metric for the progress on this refactoring.

For functions where this is not yet feasible, common reasons are:
* need access to the `uiState . uiMenu`
    * In particular, `toggleModal`/`openModal` need access to the `Menu` for use by `generateModal`.
        * And more specifically, for the "win", "lose", and "quit" modals, either to determine if it is `NoMenu` or to determine which scenario comes next.
* need access to `uiState . uiDebugOptions`
* need access to the `ScenarioCollection` (by way of `normalizeScenarioPath`) for:
    * `saveScenarioInfoOnFinishNocheat`
    * `saveScenarioInfoOnQuit`
* Need access to `uiState . uiAchievements` for:
    * `runGameTick`
 
This exercise is informative as to which record members we may want to re-organize.