Fix robot modal memory leak by only rendering modal when it is displayed (#2379)
Fixes #2370 .  Only rerender the robots modal dialog when it is actively being displayed.  This does indeed seem to get rid of the memory leak; here's a heap profile with this fix included:

![robotDisplay-heap-fixed](https://github.com/user-attachments/assets/55913af0-a46e-4140-b6e6-1c2bffddabce)

The only downside is that it now takes a noticeable fraction of a second for the robots dialog to be updated each time it is displayed.  However, I consider this a much better alternative to a memory leak.  I suppose we could try to update the robots modal immediately when it is toggled on, but personally I'd rather leave that as a future improvement.