Single authoritative source for ScenarioInfo (#2385)
Currently, the `ScenarioInfo` records for each `Scenario` are stored in two places:
* The list inside `RuntimeInfo`
* The `Menu`

This has been the source of multiple bugs: see https://github.com/swarm-game/swarm/pull/1243#pullrequestreview-1411539743, https://github.com/swarm-game/swarm/issues/1064#issuecomment-1410915316

In this PR we consolidate on the `RuntimeState` as the single authoritative source for this information and retrieve it from there at all sites that used to reference the `Menu` (in the UI).

# Implementation details

We use the `ScenarioCollection` type to represent the hierarchical structure of scenarios.  Until now this has included `ScenarioInfo` for each leaf node.  We want to re-use this type for the UI menu, but without duplicating each `ScenarioInfo`.  Therefore, we parameterize `ScenarioCollection` (and likewise `ScenarioItem`/`ScenarioInfoPair`) with a payload, which can be either a `ScenarioInfo` (for the authoritative information source) or a `ScenarioPath` (for referencing the authoritative source).

## Conversion

We can "downgrade" a `ScenarioCollection ScenarioInfo` to a `ScenarioCollection ScenarioPath` via its functor instance:
```
fmap (ScenarioPath . view scenarioPath)
```

From a `ScenarioCollection ScenarioPath`, we can recover the `ScenarioInfo` using the `loadScenarioInfoFromPath` function.

# Impact

Although the refactoring touches all type signatures involving `ScenarioInfo`, much complexity was eliminated, including:
* Deletion of the `mkNewGameMenu` function
* Removed most of the code in the `saveScenarioInfoOnQuit` function
* Eliminates a reload from disk in `constructAppState`

# Test procedure

1. Remove tutorial progress:
    ```rm ~/.local/share/swarm/saves/Tutorials_*```
2. Start the game:
    ```./scripts/play.sh```
4. Select "New game".  Observe that no tutorials show indication of progress.
5. Start the first tutorial
6. Use <kbd>Ctrl</kbd>+<kbd>q</kbd> to exit the tutorial back to the Tutorials menu
7. Observe that the tutorial entry shows an indication of progress.