Fix minor space leak in Swarm.Game.Step module
Using info table profiling we can find a small space leak in the
stepRobot function in the Swarm.Game.Step module. We can trace that
space leak back to the insert function in the Swarm.Util.WindowedCounter
module. Making the application of the Just constructor strict in that
function fixes the space leak.