Fix space leak in ActivityCounts by replacing Data.Map with Data.Map.Strict
Using info table profiling we can find a space leak that points to
the derived Ord instance for Const. Tracking the problem down we can
eventually see that the space leak originates from a use of a non-strict
Map in the ActivityCounts record. Replacing the Map with a strict one
appears to fix the space leak.