Don't redraw UI after running an animation update
This change improves performance by not redrawing the UI immediately after receiving
an update to the popup animation. The change in the popup notification
will instead be redrawn after the application processes the next frame.
Doing so will make the popup animation less smooth when
the game is struggling to keep up with Frame events. However, redrawing
the UI after processing each animation update results in redrawing the
UI potentially much more often, which hurt performance.