Use Brick.Animation to implement popup notifications (#2454)
This PR refactors the code to use `Brick.Animation` to implement notification popups. The only user-facing change this PR makes is making it so that notification popups slide in more quickly, last longer at full size, then slide out more quickly.

As part of the implementation, this PR adds a field to `AppState` to store the animation manager and adds a field to `ProgressionState` to store the state of the popup animation (including the animation itself if it is running). 

Closes #2422