Optional `chars` palette field to specify auto-generated character entities (#2494)
Adds an optional `chars` field to the `palette` specification of a world, which expects a string like `A-Za-f0-9!?` specifying a set of characters that can be used to stand for auto-generated default entities that just look like themselves.  Can be used to write text or create strictly cosmetic designs in a world.

Closes #2193 .  In the immediate future, I plan to use it for https://github.com/swarm-game/swarm/pull/2492#discussion_r2153283183 .