refactor narrowToPosition to use pathToPosition (#2473)
When getting the lsp response for hover we need to narrow the syntax tree to by a given position. This change refactors that function to do so by building a list of syntax terms and then selecting the last. This isn't useful for Hover but will be used when we need to jump to position within the tree.