Optimize structure composition (#2487)
This PR changes the structure assembly logic to build a graph of structure references, topologically sort them while checking for cycles, then assembles the structure, caching the already-merged structures in a HashMap so that we don't have to re-assemble structures.

Because two different structures in the tree of structure definitions may share the same name, we use the path of structure names from a named structure to the root as a unique identifier.

This PR resolves #1611