Move view center state to internal module (#2520)
This is a refactor before I add watch for robots (#2518).
If this is approved, I am going to move Robots
and lenses out to an internal module.

Why?

Because there is documentation for some members
and for public lenses, but I want it to be consistent.
The documentation should be on the lenses, but
some of those lenses will be only in the Internal
module and used qualified in `Robot.hs`.

Also I want to re-export modules, so that when given
a choice where to import from it is always from
the shorter module. In the perfect world, I want to
import `Swarm.Game` and get the functions that 
I actually need.