Get rid of a use of `Map.!` in `Swarm.Language.Typecheck` (#2532)
Towards #1494 .  We checked that the keys of the map representing a record type and the keys of the map representing a record literal were the same, and then traversed over the record literal, doing a lookup into the record type for each.  Since we already checked that the maps have the same keys, these lookups were guaranteed to succeed, but I still didn't like the use of `Map.!`.  This PR first uses `intersectionWith` to combine the two maps into a single map of pairs, and then traverses over the resulting map.  This accomplishes the same thing but without the use of any partial functions.