Set max frame rate to 35 FPS (#2558)
For a long time the max frame rate has been arbitrarily set to 30 FPS.  (The reason we have a max frame rate set at all is so that the game doesn't just eat 100% of your CPU.)  However, since the default speed is 16 ticks/s, doubling it results in 32 ticks/s, which at a frame rate of 30 FPS will run just under one frame per tick.  i.e. if you speed up the game just a bit, there will necessarily start being dropped ticks that are not rendered as a frame.  This seems unfortunate, so I propose upping the max frame rate slightly to better correspond to the power-of-two tick rates.  Setting the FPS to 32 exactly will probably still result in occasional dropped ticks (typically the actual frame rate hovers slightly below the maximum), so I propose setting it to 35.