Tear the World asunder (#2568)
* split World into smaller modules and re-export
* also re-export `Coords` for convenience
* split from #2555 
* depends on #2567 

The idea is to have Tile and Function modules separate from World, instead of having to scroll past them.

Then World functions get split into a Pure interface and a monadic/stateful one that should match (some are missing) and be preferred as it will implicitly use cache (it is currently not very used). 

In the next PR, the monadic interface will be improved with metrics...