Discover scenario integration tests by default instead of listing them all manually (#2599)
Previously, we had to manually list every single scenario that should be tested.  This meant in particular that if anyone created a new scenario and simply forgot to add it to the list, it would not be tested.  This happened several times in the past that I know of, and through making this PR itself we found several additional scenarios that were not listed but had a failing solution!

This PR changes the system for testing scenario solutions.  By default, every scenario in `data/scenarios` will be tested with default parameters, except: (1) those in `_Validation` are excluded; (2) there is an explicit list of scenarios that need more time; (3) there is an explicit list of scenarios with custom validation code; (4) there is an explicit list of scenarios which are expected to fail; (5) there is an explicit list of scenarios that do not have a solution.  Anything in the explicit lists overrides the default tests.  But this means that if you simply add a scenario to `data/scenarios`, it will automatically be tested in CI.  If it does not have a solution, or the solution fails, CI will fail.  You can then choose to add it to one of the explicit lists to exclude it or give it a custom configuration as necessary.

@noahyor did most of the initial coding for this PR and I helped finish it.

Closes #2397.

Note, I also want to split up `Main` into multiple modules, but I will leave that for another PR, to make the reviewing easier.