Fix `Testing/687-watch-command` scenario (#2607)
This scenario was not listed in `test/integration/Main`.  Therefore, it was not being tested.  Therefore, it bitrotted.  In particular, it was not updated when the semantics of `watch` changed in #2538. The `base` used to wake up when the `planter` robot planted a tree.  However, with the change to `watch`, the `base` woke up as soon as the `planter` moved into the cell, then immediately moved to try to `grab` the `tree`, which had not been planted yet.

The solution is to use the first `watch` to wait for the planter, then immediately execute a second `watch` and go to sleep again to wait for the tree to be planted.

Fixes #2601.