Extract world file names in a sane way (#2611)
When loading world descriptions from `data/worlds`, we extract the base name of each `foo.world` file to serve as a key into the `WorldMap`.  However, this was done in an insane, roundabout way: taking the path of the directory used to load the world files and a world file path, splitting both into components, and dropping a number of components from the second equal to the number of components in the first.  However, when the two ended up normalized differently, (i.e. the first contained an extra `/./`) this did not work.

This PR simplifies the code to just call the standard `takeFileName` function.

Fixes #2610.