fix: Use Named Fields for GrowthTime (#2636)
Fixes #2626 
Hello! This is my first try at a PR on a Haskell project. My Haskell skills aren't that great, so any constructive criticism is welcome!
Also I couldn't get HLS to work so I am not sure if I found all the places where this was being used.
Another thing to call-out is that I am not too sure about the deriving strategies I used and if they are correct.