Add `swarm check` for parsing + checking `.sw` files from command-line (#2640)
Add a `swarm check` subcommand, which takes the name of a `.sw` file, or else reads from standard input if given the `--stdin` flag, similarly to the `format` subcommand.  Prints `OK.` if the file parses + typechecks, otherwise prints out the same error message that would be printed at the REPL.  Also yields an appropriate exit code, so `swarm check` could be used as part of some automated shell process.

Closes #2284.