Add `--v0.7` flag to `swarm format` and get rid of `--v0.6` conversion (#2632)
The only breaking change from 0.7 to 0.8 is that the type of `atomic` and `instant` changed (see #2552).  In particular their argument type changed from `Cmd a` to `{Cmd a}`.  Therefore, `swarm format --v0.7` finds any application of `atomic` or `instant` and wraps the argument in delay brackets.

```
$ echo 'atomic move' | cabal run swarm:exe:swarm -- format --stdin
atomic move
$ echo 'atomic move' | cabal run swarm:exe:swarm -- format --stdin --v0.7
atomic {move}
```

This PR also rips out the code to convert 0.6 -> 0.7; according to convention at the moment, we only include code to upgrade one version at a time.