Close #1149 as wontfix (#2642)
Delete a TODO comment linking to #1149.  The idea was to use some generic Unicode tools to automatically create a pool of viable map palette characters (e.g. ones that do not have a wide display). However, as explained in the comments at #1149, it appears that there is really no way to do that.

After merging this PR, we should close the relevant issue as wontfix.